Resolved issue with multiple inheritance.
We had two classes, Threadable and VideoWriter that implemented a context
manager, as well as open/close methods. My understanding of Python
resolution order is that it should have looked left-to-right, however
something was causing the Threadable open/close or __enter/__exit methods
not to run. The result is the video writer never starts the thread!
When you call write(), an error is thrown...

There's already a test for this, but somehow the functionality didn't
work for me. I tested this in my project, and it seemed to work great!